Integrate in existing hydrus framework
