+++ hydrus/data/crud.py
@@ -35,8 +35,7 @@ import re
 from sqlalchemy.orm import with_polymorphic
 from sqlalchemy import exists
 from sqlalchemy.orm.exc import NoResultFound
-from hydrus.data.db_models import (Graph, BaseProperty, RDFClass, Instance,
-                                   Terminal, GraphIAC, GraphIIT, GraphIII, Modification)
+from hydrus.data.db_models import Modification
 from hydrus.data.exceptions import (
     ClassNotFound,
     InstanceExists,
@@ -54,14 +53,18 @@ from hydrus.data.crud_helpers import (
     get_rdf_class,
     get_data_iac_iii_iit,
     add_prop_name_to_object,
-    get_instance_before_delete,
-    get_all_filtered_instances)
+    get_instance_before_delete)
 # from sqlalchemy.orm.session import Session
 from sqlalchemy.orm.scoping import scoped_session
 from typing import Dict, Optional, Any, List
 
-triples = with_polymorphic(Graph, '*')
-properties = with_polymorphic(BaseProperty, "*")
+from hydrus.data.resource_based_classes import (
+    get_object,
+    insert_object,
+    update_object,
+    delete_object,
+    get_all_filtered_instances
+)
 
 
 def get(id_: str, type_: str, api_name: str, session: scoped_session,
@@ -80,12 +83,11 @@ def get(id_: str, type_: str, api_name: str, session: scoped_session,
         InstanceNotFound: If no Instance of the 'type_` class if found.
 
     """
-    object_template = {
-        "@type": "",
-    }  # type: Dict[str, Any]
-    rdf_class = get_rdf_class(session, type_)
-    object_template = add_prop_name_to_object(session, id_, object_template, rdf_class)
-    object_template["@type"] = rdf_class.name
+    query_info = {
+        "@type": type_,
+        "id_": id_
+    }
+    object_template = get_object(query_info, session)
     if path is not None:
         object_template["@id"] = f"/{api_name}/{path}Collection/{id_}"
     else:
@@ -113,66 +115,11 @@ def insert(object_: Dict[str, Any], session: scoped_session, link_props: Dict[st
             not an Instance property
         NotAbstractProperty: If any property of `object_` is a
             valid/defined RDFClass but is not a dictionary neither an Abstract Property
-
     """
-    rdf_class = None
-    instance = None
-    type_ = object_["@type"]
-    # Check for class in the beginning
-    rdf_class = get_rdf_class(session, type_)
-    if id_ is not None and session.query(exists().where(Instance.id == id_)).scalar():
-        raise InstanceExists(type_=rdf_class.name, id_=id_)
-    elif id_ is not None:
-        instance = Instance(id=id_, type_=rdf_class.id)
-    else:
-        instance = Instance(type_=rdf_class.id)
-    session.add(instance)
-    session.flush()
-
-    for prop_name in object_:
-
-        if prop_name not in ["@type", "@context"]:
-            try:
-                property_ = session.query(properties).filter(
-                    properties.name == prop_name).one()
-            except NoResultFound:
-                # Adds new Property
-                session.close()
-                raise PropertyNotFound(type_=prop_name)
-            # For insertion in III through link
-            if prop_name in link_props:
-                try:
-                    insert_iii_with_link(instance.id, property_,
-                                         link_props[prop_name], session)
-                except (NotInstanceProperty, InstanceNotFound, ClassNotFound):
-                    raise
-                continue
-            # For insertion in III
-            if isinstance(object_[prop_name], dict):
-                try:
-                    insert_iii(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotInstanceProperty:
-                    raise
-            # For insertion in IAC
-            elif session.query(exists().where(RDFClass.name == str(object_[prop_name]))).scalar() \
-                    and property_.type_ == "PROPERTY" or property_.type_ == "ABSTRACT":
-                try:
-                    insert_iac(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotAbstractProperty:
-                    raise
-
-            # For insertion in IIT
-            else:
-                try:
-                    insert_iit(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotInstanceProperty:
-                    raise
-
-    session.commit()
-    return instance.id
+    if id_ is not None:
+        object_['id'] = id_
+    inserted_object_id = insert_object(object_, session)
+    return inserted_object_id
 
 
 def insert_multiple(objects_: List[Dict[str,
@@ -237,27 +184,11 @@ def delete(id_: str, type_: str, session: scoped_session) -> None:
         InstanceNotFound: If no instace of type `type_` with id `id_` exists.
 
     """
-    instance = get_instance_before_delete(session, id_, type_)
-    data_IAC, data_III, data_IIT = get_data_iac_iii_iit(session, id_)
-    data = data_III + data_IIT + data_IAC
-    for item in data:
-        session.delete(item)
-
-    for data in data_IIT:
-        terminal = session.query(Terminal).filter(
-            Terminal.id == data.object_).one()
-        session.delete(terminal)
-
-    for data in data_III:
-        III_instance = session.query(Instance).filter(
-            Instance.id == data.object_).one()
-        III_instance_type = session.query(RDFClass).filter(
-            RDFClass.id == III_instance.type_).one()
-        # Get the III object type_
-        delete(III_instance.id, III_instance_type.name, session=session)
-
-    session.delete(instance)
-    session.commit()
+    query_info = {
+        "@type": type_,
+        "id_": id_
+    }
+    delete_object(query_info, session)
 
 
 def delete_multiple(
@@ -299,21 +230,12 @@ def update(id_: str,
     :param path: endpoint
     :return: id of updated object
     """
-    # Keep the object as fail safe
-    instance = get(id_=id_, type_=type_, session=session, api_name=api_name)
-    instance.pop("@id")
-    # Delete the old object
-    delete(id_=id_, type_=type_, session=session)
-    # Try inserting new object
-    try:
-        insert(object_=object_, id_=id_, link_props=link_props, session=session)
-    except (ClassNotFound, InstanceExists, PropertyNotFound) as e:
-        # Put old object back
-        insert(object_=instance, id_=id_, link_props=link_props, session=session)
-        raise e
-
-    get(id_=id_, type_=type_, session=session, api_name=api_name, path=path)
-    return id_
+    query_info = {
+        "@type": type_,
+        "id_": id_
+    }
+    updated_object_id = update_object(object_, query_info, session)
+    return updated_object_id
 
 
 def get_collection(API_NAME: str,
@@ -570,152 +492,3 @@ def get_modification_table_diff(session: scoped_session,
         }
         list_of_modification_records.append(modification_record)
     return list_of_modification_records
-
-
-def insert_iii(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIII:
-    """
-    Insert a GraphIII triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIII triple.
-
-    :raises: NotInstanceProperty
-    """
-    instance_id = insert(object_[prop_name], session=session)
-    instance_object = session.query(Instance).filter(
-        Instance.id == instance_id).one()
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        session.add(property_)
-        triple = GraphIII(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=instance_object.id)
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotInstanceProperty(type_=prop_name)
-
-
-def insert_iac(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIAC:
-    """
-    Insert a GraphIAC triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIAC triple.
-    """
-    if property_.type_ == "PROPERTY" or property_.type_ == "ABSTRACT":
-        property_.type_ = "ABSTRACT"
-        session.add(property_)
-        class_ = session.query(RDFClass).filter(
-            RDFClass.name == object_[prop_name]).one()
-        triple = GraphIAC(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=class_.id)
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotAbstractProperty(type_=prop_name)
-
-
-def insert_iit(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIIT:
-    """
-    Insert a GraphIIT triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIIT triple.
-
-    :raises: NotInstanceProperty
-    """
-    terminal = Terminal(value=object_[prop_name])
-    session.add(terminal)
-    session.flush()  # Assigns ID without committing
-
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        session.add(property_)
-        triple = GraphIIT(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=terminal.id)
-        # Add things directly to session, if anything fails whole
-        # transaction is aborted
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotInstanceProperty(type_=prop_name)
-
-
-def insert_iii_with_link(instance_id: str, property_: BaseProperty,
-                         property_value: str, session: scoped_session):
-    """
-    Inserts GraphIII triple to store a relation defined with hydra:Link.
-    :param instance_id:  Id of the instance being inserted
-    :param property_: Property being used as predicate in the new triple.
-    :param property_value: Value of the property being inserted.
-    :param session: sqlalchemy session
-    :return:
-    """
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        # If value matches with the regex then value is an id and link is to an
-        # instance of a collection class otherwise value is a class_type and link
-        # is to a non collection class.
-        regex = r'[a-z0-9]{8}-([a-z0-9]{4}-){3}[a-z0-9]{12}'
-        matchObj = re.match(regex, property_value)
-        # Link is to an instance of a collection class
-        if matchObj:
-            try:
-                nested_instance = session.query(Instance).filter(
-                    Instance.id == property_value).one()
-            except NoResultFound:
-                raise InstanceNotFound(id_=property_value, type_="")
-            triple = GraphIII(
-                subject=instance_id,
-                predicate=property_.id,
-                object_=nested_instance.id)
-            session.add(triple)
-            return triple
-        # Link is to a non collection, single instance class
-        else:
-            try:
-                nested_rdf_class = session.query(RDFClass).filter(
-                    RDFClass.name == property_value).one()
-            except NoResultFound:
-                raise ClassNotFound(type_=property_value)
-            try:
-                nested_instance = session.query(Instance).filter(
-                    Instance.type_ == nested_rdf_class.id).all()[-1]
-            except (NoResultFound, IndexError, ValueError):
-                raise InstanceNotFound(type_=nested_rdf_class.name)
-            triple = GraphIII(
-                subject=instance_id,
-                predicate=property_.id,
-                object_=nested_instance.id)
-            session.add(triple)
-            return triple
-    else:
-        session.close()
-        raise NotInstanceProperty(type_=property_.name)
