+++ hydrus/data/db_models.py
@@ -1,231 +1,139 @@
 """Models for Hydra Classes."""
 
-from sqlalchemy import create_engine, ForeignKey
-from sqlalchemy.sql import func
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy import Column, Integer, String, DateTime
-from sqlalchemy.sql import func
-from typing import Any
 import datetime
 import uuid
+from sqlite3 import Connection as SQLite3Connection
+from typing import Any
+
+from sqlalchemy import (
+    Column,
+    DateTime,
+    ForeignKey,
+    Integer,
+    String,
+    create_engine,
+    event,
+)
+from sqlalchemy.engine import Engine
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.sql import func
+
 # from hydrus.settings import DB_URL
 
-engine = create_engine('sqlite:///database.db')
+engine = create_engine("sqlite:///database.db")
 
 Base = declarative_base()  # type: Any
 
 EXPIRY_TIME = 2700  # unit: seconds
 
 
-class RDFClass(Base):
-    """Model for Classes.
-    Classes are RDF-OWL or RDF-HYDRA classes.
+# required for enforcing foreign key constraints through SQLite
+@event.listens_for(Engine, "connect")
+def _set_sqlite_pragma(dbapi_connection, connection_record):
+    if isinstance(dbapi_connection, SQLite3Connection):
+        cursor = dbapi_connection.cursor()
+        cursor.execute("PRAGMA foreign_keys=ON;")
+        cursor.close()
+
+
+class Resource:
+    """
+    Class for interacting with any resource(class or collection) in the ApiDoc
     """
 
-    __tablename__ = "classes"
+    # TODO: Look for a better way to store the sql alchemy class for each resource
+    all_database_classes = {}
 
-    id = Column(
-        String,
-        default=lambda: str(
-            uuid.uuid4()),
-        unique=True,
-        primary_key=True)
-    name = Column(String, unique=True)
+    def __init__(self, resource):
+        self._resource = resource
 
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<id='{self.id}', name='{self.name}'>"
+    @property
+    def name(self):
+        """Return the name of the resource from it's "@id"""
+        # split the classname at "vocab:" to get the class name
+        return self._resource["@id"].split("vocab:")[1]
 
+    @property
+    def resource(self):
+        """Return the resource dict"""
+        return self._resource
 
-class Instance(Base):
-    """Model for Object/Resource.
-    Instances are instances of some kind/classes that are served through the API.
-    """
+    @property
+    def supported_properties(self):
+        """
+        Return all the properties in "supportedProperty" property
+        of that resource.
+        """
+        return self._resource["supportedProperty"]
 
-    __tablename__ = "instances"
+    def get_attr_dict(self):
+        """
+        Return the attribute dictionary necessary for
+        creating that resource's table at runtime
+        """
+        # initialize the attribute dict with "__tablename__"
+        # and a "id" column which will act as primary key
+        # for instances of that table
+        attr_dict = {
+            "__tablename__": self.name,
+            "id": Column(
+                String,
+                default=lambda: str(uuid.uuid4()),
+                unique=True,
+                primary_key=True,
+            ),
+        }
+        for supported_property in self.supported_properties:
+            title = supported_property["title"]
+            link = supported_property["property"]
+            if type(link) is not dict:
+                if "vocab:" in link:
+                    # if vocab: is in the link, it implies that the link is pointing to
+                    # another resource in the same ApiDoc, hence make it a Foreign Key
+                    # to that resource table
+                    foreign_table_name = link.split("vocab:")[1]
+                    attr_dict[title] = Resource.foreign_key_column(
+                        foreign_table_name
+                    )
 
-    id = Column(
-        String,
-        default=lambda: str(
-            uuid.uuid4()),
-        unique=True,
-        primary_key=True)
-    type_ = Column(String, ForeignKey("classes.id"), nullable=True)
-    created = Column('created', DateTime, default=func.now())
-    last_modified = Column('last_modified', DateTime, onupdate=func.now())
+                else:
+                    attr_dict[title] = Column(String)
+            else:
+                # if the supported property has "property" attribute of @type "hydra:link"
+                if "vocab:" in link["range"]:
+                    # if vocab: is in the link["range"], it implies that the link is pointing to
+                    # another resource in the same ApiDoc, hence make it a Foreign Key
+                    # to that resource table
+                    foreign_table_name = link["range"].split("vocab:")[1]
+                    attr_dict[title] = Resource.foreign_key_column(
+                        foreign_table_name
+                    )
+                else:
+                    attr_dict[title] = Column(String)
 
+        return attr_dict
 
-class BaseProperty(Base):
-    """Model for Basic Property."""
+    @staticmethod
+    def foreign_key_column(foreign_table_name):
+        """
+        Return a sqlalchemy column which will act as
+        a foreign key to given tablename.
+        """
 
-    __tablename__ = "property"
+        return Column(
+            String,
+            ForeignKey(
+                f"{foreign_table_name}.id",
+                ondelete="CASCADE",
+                onupdate="CASCADE",
+            ),
+        )
 
-    id = Column(
-        String,
-        default=lambda: str(
-            uuid.uuid4()),
-        unique=True,
-        primary_key=True)
-    name = Column(String, unique=True, nullable=False)
-    type_ = Column(String)
-
-    __mapper_args__ = {
-        'polymorphic_on': type_,
-        'polymorphic_identity': 'PROPERTY'
-    }
-
-
-class InstanceProperty(BaseProperty):
-    """Model for Instance Properties.
-    Instance Properties are properties that are used as predicate when the subject is an Instance.
-    >>> prop1 = BaseProperty('hasWeight')
-    >>> prop2 = BaseProperty('hasCost')
-    """
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'INSTANCE'
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<id='{self.id}', name='{self.name}', type='{self.type_}'>"
-
-
-class AbstractProperty(BaseProperty):
-    """Model for Abstract Properties.
-    Abstract Properties are properties that are used as predicate between two RDF-OWL classes.
-    >>> prop1 = BaseProperty('hasWeight')
-    >>> prop2 = BaseProperty('hasCost')
-    Example of a triple:
-     RDFClass('A') Property('isSubSystemOf') RDFClass('B')
-    """
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'ABSTRACT'
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<id='{self.id}', name='{self.name}', type='{self.type_}'>"
-
-
-class Terminal(Base):
-    """Model for Terminals.
-    Terminals are numbers or string that can be referenced by a Property. They can be only
-     objects in a triple.
-    >>> t = Terminal(value=85, unit='cubic centimeters')
-    >>> t1 = Terminal(value='Cubesat', unit='standard')
-    """
-
-    __tablename__ = "terminals"
-
-    id = Column(
-        String,
-        default=lambda: str(
-            uuid.uuid4()),
-        unique=True,
-        primary_key=True)
-    value = Column(String)
-    unit = Column(String)
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<id='{self.id}', value='{self.value}', unit='{self.unit}'>"
-
-
-class Graph(Base):
-    """Model for a graph that store triples of instance from the other models
-         to map relationships."""
-
-    __tablename__ = "graph"
-
-    id = Column(
-        String,
-        default=lambda: str(
-            uuid.uuid4()),
-        unique=True,
-        primary_key=True)
-    type = Column(String)
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'graph',
-        'polymorphic_on': type
-    }
-
-
-class GraphCAC(Graph):
-    """Graph model for Class >> AbstractProperty >> Class."""
-
-    __tablename__ = 'graphcac'
-    id = Column(String, ForeignKey('graph.id'), primary_key=True)
-    subject = Column(String, ForeignKey("classes.id"))
-    predicate = Column(String, ForeignKey("property.id"))
-    object_ = Column(String, ForeignKey("classes.id"))
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'graphcac',
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<subject='{self.subject}', predicate='{self.predicate}', object_='{self.object_}'>"
-
-
-class GraphIAC(Graph):
-    """Graph model for Instance >> AbstractProperty >> Class."""
-
-    __tablename__ = 'graphiac'
-    id = Column(String, ForeignKey('graph.id'), primary_key=True, default=lambda: str(
-        uuid.uuid4()))
-    subject = Column(String, ForeignKey("instances.id"))
-    predicate = Column(String, ForeignKey("property.id"))
-    object_ = Column(String, ForeignKey("classes.id"))
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'graphiac',
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<subject='{self.subject}', predicate='{self.predicate}', object_='{self.object_}'>"
-
-
-class GraphIII(Graph):
-    """Graph model for Instance >> InstanceProperty >> Instance."""
-
-    __tablename__ = 'graphiii'
-    id = Column(String, ForeignKey('graph.id'), primary_key=True, default=lambda: str(
-        uuid.uuid4()))
-    subject = Column(String, ForeignKey("instances.id"))
-    predicate = Column(String, ForeignKey("property.id"))
-    object_ = Column(String, ForeignKey("instances.id"))
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'graphiii',
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<subject='{self.subject}', predicate='{self.predicate}', object_='{self.object_}'>"
-
-
-class GraphIIT(Graph):
-    """Graph model for Instance >> InstanceProperty >> Terminal."""
-
-    __tablename__ = 'graphiit'
-    id = Column(String, ForeignKey('graph.id'), primary_key=True, default=lambda: str(
-        uuid.uuid4()))
-    subject = Column(String, ForeignKey("instances.id"))
-    predicate = Column(String, ForeignKey("property.id"))
-    object_ = Column(String, ForeignKey("terminals.id"))
-
-    __mapper_args__ = {
-        'polymorphic_identity': 'graphiit',
-    }
-
-    def __repr__(self) -> str:
-        """Verbose object name."""
-        return f"<subject='{self.subject}', predicate='{self.predicate}', object_='{self.object_}'>"
+    def make_db_table(self):
+        """Generate the sqlalchemy table class for that resource"""
+        self.table_class = type(self.name, (Base,), self.get_attr_dict())
+        # add that class to dict for future lookups
+        Resource.all_database_classes[self.name] = self.table_class
 
 
 class Modification(Base):
@@ -233,10 +141,7 @@ class Modification(Base):
 
     __tablename__ = "modifications"
 
-    job_id = Column(
-        Integer,
-        primary_key=True,
-        autoincrement=True)
+    job_id = Column(Integer, primary_key=True, autoincrement=True)
     method = Column(String)
     resource_url = Column(String)
 
@@ -256,14 +161,18 @@ class Token(Base):
     __tablename__ = "tokens"
     id = Column(
         String,
-        default=lambda: str(
-            uuid.uuid4()),
+        default=lambda: str(uuid.uuid4()),
         unique=True,
-        primary_key=True)
-    user_id = Column(Integer, ForeignKey('user.id'))
+        primary_key=True,
+    )
+    user_id = Column(Integer, ForeignKey("user.id"))
     timestamp = Column(DateTime, default=func.now())
-    expiry = Column('expiry', DateTime, default=datetime.datetime.utcnow() +
-                    datetime.timedelta(seconds=EXPIRY_TIME))
+    expiry = Column(
+        "expiry",
+        DateTime,
+        default=datetime.datetime.utcnow()
+        + datetime.timedelta(seconds=EXPIRY_TIME),
+    )
 
     def is_valid(self):
         if self.expiry > datetime.datetime.utcnow():
@@ -277,6 +186,12 @@ class Nonce(Base):
     __tablename__ = "nonce"
     id = Column(String, primary_key=True)
     timestamp = Column(DateTime)
+
+
+def create_database_tables(classes):
+    for single_class in classes:
+        resource = Resource(single_class)
+        resource.make_db_table()
 
 
 if __name__ == "__main__":
