+++ hydrus/data/doc_parse.py
@@ -1,7 +1,6 @@
 """Parser for Hydra APIDocumentation creates Classes and Properties."""
 from sqlalchemy import exists
 
-from hydrus.data.db_models import RDFClass, BaseProperty
 from typing import Any, Dict, List, Set, Optional
 from sqlalchemy.orm.session import Session
 from sqlalchemy.orm import scoped_session
