+++ hydrus/item_collection_helpers.py
@@ -14,7 +14,8 @@ from hydrus.data.exceptions import (
     InstanceNotFound,
     PageNotFound,
     InvalidSearchParameter,
-    OffsetOutOfRange)
+    OffsetOutOfRange,
+    DatabaseConstraintError)
 
 from hydrus.helpers import (
     set_response_headers,
@@ -132,7 +133,7 @@ def item_collection_put_response(path: str) -> Response:
                 return set_response_headers(
                     jsonify(status.generate()), headers=headers_,
                     status_code=status.code)
-            except (ClassNotFound, InstanceExists, PropertyNotFound) as e:
+            except (ClassNotFound, InstanceExists, PropertyNotFound, DatabaseConstraintError) as e:
                 error = e.get_HTTP()
                 return error_response(error)
         else:
